pyproject.toml
[project]
name = "bot-curriculum"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.86.0",
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-anthropic>=1.4.0",
    "liteparse>=1.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "requests>=2.33.0",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
]
